requirements.txt
aiowebsocket==1.0.0.dev2
anyio==4.2.0
certifi==2023.11.17
filetype==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
pillow==10.2.0
sniffio==1.3.0
